+++ examples/base/headless/src/App.tsx
@@ -18,7 +18,7 @@ const App: React.FC = () => {
                     edit: PostEdit,
                 },
             ]}
-            disableTelemetry={true}
+            options={{ disableTelemetry: true }}
         />
     );
 };