+++ examples/blog/ecommerce/pages/_app.tsx
@@ -31,10 +31,10 @@ function MyApp({ Component, pageProps }: AppProps): JSX.Element {
                 dataProvider={dataProvider}
                 resources={[{ name: "products" }]}
                 Layout={Layout}
-                reactQueryDevtoolConfig={{
-                    position: "bottom-left",
+                options={{
+                    reactQuery: { devtoolConfig: { position: "bottom-left" } },
+                    disableTelemetry: true,
                 }}
-                disableTelemetry={true}
             >
                 <ChakraProvider>
                     <Component {...pageProps} />