+++ examples/calendar/src/App.tsx
@@ -20,7 +20,7 @@ const App: React.FC = () => {
                 },
             ]}
             Layout={Layout}
-            disableTelemetry={true}
+            options={{ disableTelemetry: true }}
         />
     );
 };