+++ examples/core/useImport/src/App.tsx
@@ -10,7 +10,7 @@ const App: React.FC = () => {
             dataProvider={dataProvider("https://api.fake-rest.refine.dev")}
             routerProvider={routerProvider}
             resources={[{ name: "posts", list: DummyList }]}
-            disableTelemetry={true}
+            options={{ disableTelemetry: true }}
         />
     );
 };