+++ examples/fineFoods/client/pages/_app.tsx
@@ -17,7 +17,7 @@ function MyApp({ Component, pageProps }: AppProps): JSX.Element {
             dataProvider={dataProvider(API_URL)}
             Layout={Layout}
             resources={[{ name: "users" }]}
-            disableTelemetry={true}
+            options={{ disableTelemetry: true }}
         >
             <Head>
                 <title>finefoods client example - refine</title>