+++ packages/antd/tsup.config.ts
@@ -1,4 +1,5 @@
 import { defineConfig } from "tsup";
+import * as fs from "fs";
 import copyStaticFiles from "esbuild-copy-static-files";
 import { NodeResolvePlugin } from "@esbuild-plugins/node-resolve";
 
@@ -9,6 +10,32 @@ export default defineConfig({
     clean: false,
     platform: "browser",
     esbuildPlugins: [
+        {
+            name: "react-remove-testids",
+            setup(build) {
+                build.onEnd(async (args) => {
+                    // data-testid regexp
+                    const regexp = /("data-testid":)(.*?)(?:(,)|(}))/gi;
+
+                    // output files with `*.js`
+                    const jsOutputFiles =
+                        args.outputFiles?.filter((el) =>
+                            el.path.endsWith(".js"),
+                        ) ?? [];
+
+                    // replace data-testid in output files
+                    for (const jsOutputFile of jsOutputFiles) {
+                        const str = new TextDecoder("utf-8").decode(
+                            jsOutputFile.contents,
+                        );
+                        const newStr = str.replace(regexp, "$4");
+                        jsOutputFile.contents = new TextEncoder().encode(
+                            newStr,
+                        );
+                    }
+                });
+            },
+        },
         copyStaticFiles({
             src: "./src/assets/styles/styles.min.css",
             dest: "./dist/styles.min.css",