+++ packages/core/src/hooks/accessControl/useCan/index.ts
@@ -34,8 +34,16 @@ export const useCan = ({
      */
     const { resource: _resource, ...paramsRest } = params ?? {};
 
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const { icon: _icon, ...restResource } = _resource ?? {};
+    /* eslint-disable @typescript-eslint/no-unused-vars */
+    const {
+        icon: _icon,
+        list: _list,
+        edit: _edit,
+        create: _create,
+        show: _show,
+        ...restResource
+    } = _resource ?? {};
+    /* eslint-enable @typescript-eslint/no-unused-vars */
 
     const queryResponse = useQuery<CanReturnType>(
         [
@@ -47,7 +55,9 @@ export const useCan = ({
             },
         ],
         // Enabled check for `can` is enough to be sure that it's defined in the query function but TS is not smart enough to know that.
-        () => can?.({ action, resource, params }) ?? { can: true },
+        () =>
+            can?.({ action, resource, params }) ??
+            Promise.resolve({ can: true }),
         {
             enabled: typeof can !== "undefined",
             ...queryOptions,