+++ packages/mui/test/jest.setup.ts
@@ -1,7 +1,12 @@
 import "@testing-library/jest-dom";
 import "@testing-library/jest-dom/extend-expect";
+import { configure } from "@testing-library/dom";
 import * as util from "util";
 // import { createMockServer } from "./dataMocks";
+
+configure({
+    asyncUtilTimeout: 10000,
+});
 
 /** Antd mocks */
 window.matchMedia = jest.fn().mockImplementation((query) => {
@@ -32,3 +37,5 @@ Object.defineProperty(window, "TextDecoder", {
     writable: true,
     value: util.TextDecoder,
 });
+
+jest.setTimeout(20000);